% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitnavg.R
\name{logitnavgcdf}
\alias{logitnavgcdf}
\title{Cumulative distribution function of the average of two logit-normal variables}
\usage{
logitnavgcdf(x, mu1, sigma1, mu2, sigma2)
}
\arguments{
\item{x}{Vector of values.}

\item{mu1}{Location parameter of the first distribution.}

\item{sigma1}{Scale parameter of the first distribution.}

\item{mu2}{Location parameter of the second distribution.}

\item{sigma2}{Scale parameter of the second distribution.}
}
\value{
Value of the cumulative distribution function.
}
\description{
Cumulative distribution function of the average of two logit-normal variables
}
