% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitnavg.R
\name{logitnavginv}
\alias{logitnavginv}
\title{Inverse cumulative distribution function of the average of two logit-normal
variables}
\usage{
logitnavginv(y, mu1, sigma1, mu2, sigma2)
}
\arguments{
\item{y}{Vector of quantiles.}

\item{mu1}{Location parameter of the first distribution.}

\item{sigma1}{Scale parameter of the first distribution.}

\item{mu2}{Location parameter of the second distribution.}

\item{sigma2}{Scale parameter of the second distribution.}
}
\value{
Value of the inverse cumulative distribution function of the average
of two logit-normal variables.
}
\description{
Inverse cumulative distribution function of the average of two logit-normal
variables
}
