% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitnavg.R
\name{logitnavgmean}
\alias{logitnavgmean}
\title{Expectation of the average of two logit-normal variables}
\usage{
logitnavgmean(mu1, sigma1, mu2, sigma2)
}
\arguments{
\item{mu1}{Location parameter of the first distribution.}

\item{sigma1}{Scale parameter of the first distribution.}

\item{mu2}{Location parameter of the second distribution.}

\item{sigma2}{Scale parameter of the second distribution.}
}
\value{

}
\description{
Expectation of the average of two logit-normal variables
}
