% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitn.R
\name{logitncdf}
\alias{logitncdf}
\title{Logit-normal cumulative distribution function}
\usage{
logitncdf(x, mu, sigma)
}
\arguments{
\item{x}{Vector of values.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter.}
}
\value{
Value of the logit-normal cumulative distribution function.
}
\description{
Logit-normal cumulative distribution function
}
\examples{
logitncdf(0.5, 1, 2)
logitncdf(c(0, 0.5), c(0, 1), 2)
}
