% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitn.R
\name{logitnconv}
\alias{logitnconv}
\title{Convolution of two logit-normal distributions}
\usage{
logitnconv(res, mu1, sigma1, mu2, sigma2)
}
\arguments{
\item{res}{Desired resolution, expressed as distribution support per data
point. Since the support is [0, 2], a resolution of 0.1 returns 21 points.}

\item{mu1}{Location parameter of the first distribution.}

\item{sigma1}{Scale parameter of the first distribution.}

\item{mu2}{Location parameter of the second distribution.}

\item{sigma2}{Scale parameter of the second distribution.}
}
\value{
A vector of values containing the density of the convolution of two
logit-normal distributions, evaluated on a grid of values between 0 and 2 at
the specified resolution.
}
\description{
Convolution of two logit-normal distributions
}
