% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitn.R
\name{logitninv}
\alias{logitninv}
\title{Logit-normal inverse cumulative distribution function}
\usage{
logitninv(p, mu, sigma)
}
\arguments{
\item{p}{Vector of quantiles.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter.}
}
\value{
Value of the logit-normal inverse cumulative distribution function.
}
\description{
Logit-normal inverse cumulative distribution function
}
\examples{
logitninv(0.5, 1, 2)
logitninv(c(0, 0.5), c(0, 1), 2)
}
