% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitn.R
\name{logitnmean}
\alias{logitnmean}
\title{Expectation of the logit-normal distribution}
\usage{
logitnmean(mu, sigma)
}
\arguments{
\item{mu}{Location parameter.}

\item{sigma}{Scale parameter.}
}
\value{
Expectation.
}
\description{
Expectation of the logit-normal distribution
}
\examples{
logitnmean(0, 0.6)  # 0.5
}
