% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitn.R
\name{logitnpdf}
\alias{logitnpdf}
\title{Logit-normal probability density function}
\usage{
logitnpdf(x, mu, sigma)
}
\arguments{
\item{x}{Vector of values.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter.}
}
\value{
Density of the logit-normal distribution.
}
\description{
Logit-normal probability density function
}
\examples{
logitnpdf(0.5, 1, 2)
logitnpdf(c(0, 0.5), c(0, 1), 2)
}
