% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micp-package.R
\docType{package}
\name{micp-package}
\alias{micp}
\alias{micp-package}
\title{micp: Bayesian mixed-effects inference on classification performance}
\description{
An R package for variational Bayesian mixed-effects inference on classification performance.
}
\keyword{internal}
