% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{repmat}
\alias{repmat}
\title{Replicate a matrix.}
\usage{
repmat(X, m = 1L, n = 1L)
}
\arguments{
\item{X}{Vector or matrix.}

\item{m}{Number of vertical replications.}

\item{n}{Number of horizontal replications.}
}
\value{
Replicated matrix.
}
\description{
Replicate a matrix.
}
