% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbicp.unb.R
\name{update.mu}
\alias{update.mu}
\title{Update the 2nd mean-field partition (mu)}
\usage{
\method{update}{mu}(data, prior, q)
}
\arguments{
\item{data}{List of `ks`, `ns`, and `m`.}

\item{prior}{List of `mu.0`, `eta.0`, `a.0`, `b.0`.}

\item{q}{List of approximate posterior moments.}
}
\value{
A revised `q` with updated `mu.mu` and `eta.mu` elements.
}
\description{
Update the 2nd mean-field partition (mu)
}
