% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbicp.unb.R
\name{update.rho}
\alias{update.rho}
\title{Update the 1st mean-field partition (rho)}
\usage{
\method{update}{rho}(data, prior, q)
}
\arguments{
\item{data}{List of `ks`, `ns`, and `m`.}

\item{prior}{List of `mu.0`, `eta.0`, `a.0`, `b.0`.}

\item{q}{List of approximate posterior moments.}
}
\value{
A revised `q` with updated `mu.rho` and `eta.rho` elements.
}
\description{
Update the 1st mean-field partition (rho)
}
