% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitn.R
\name{logitnpdf}
\alias{logitnpdf}
\title{Logit-normal probability density function}
\usage{
logitnpdf(x, mu, sigma)
}
\arguments{
\item{sigma}{}
}
\value{

}
\description{
Logit-normal probability density function
}
\examples{
logitnpdf(0.5, 1, 2)
}
