% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micp.stats.R
\name{micp.stats}
\alias{micp.stats}
\title{Mixed-effects inference on classification performance}
\usage{
micp.stats(ks, ns)
}
\arguments{
\item{ks}{When inferring on accuracies: m-sized vector of number of correct
predictions (in each subject). When inferring on balanced accuracies: 2xm
matrix of correctly predicted positive trials (first row) and corectly
predicted negative trials (second row).}

\item{ns}{When inferring on accuracies: m-sized vector of total number of
trials (in each subject). When inferring on balanced accuracies: 2xm matrix
of total number of positive (first row) and negative (second row) trials.}
}
\value{
A list with the following fields: mu, p, ci, and stats.

  mu: Posterior mean of the population mean accuracy or balanced accuracy.
    This is the expected performance of the classifier at the group level.

  p: Posterior infraliminal probability of the population mean. This is
    the posterior belief that the classifier did not operate above chance
    (50%). A small infraliminal probability represents evidence for
    above-chance performance.

  ci: Posterior 95% credible interval of the population mean. This interval
    can be used to show error bars around mu.

  stats: Additional return values, depending on the selected model. See
    individual inference functions for details.
}
\description{
Mixed-effects inference on classification performance
}
\examples{
  # Accuracy:
  ks <- c(19, 41, 15, 39, 39)
  ns <- c(45, 51, 20, 46, 58)
  results1 <- micp::micp.stats(ks, ns)
  print(results1)

  # Balanced accuracy:
  ks <- rbind(c(19, 41, 15, 39, 39), c(41, 46, 43, 48, 37))
  ns <- rbind(c(45, 51, 20, 46, 58), c(55, 49, 80, 54, 42))
  results2 <- micp::micp.stats(ks, ns)
  print(results2)
}
\references{
K.H. Brodersen, J. Daunizeau, C. Mathys, J.R. Chumbley, J.M. Buhmann, &
  K.E. Stephan (2013). Variational Bayesian mixed-effects inference for
  classification studies. NeuroImage (in press).
  doi:10.1016/j.neuroimage.2013.03.008.

  K.H. Brodersen, C. Mathys, J.R. Chumbley, J. Daunizeau, C.S. Ong, J.M.
  Buhmann, & K.E. Stephan (2012). Bayesian mixed-effects inference on
  classification performance in hierarchical datsets. Journal of Machine
  Learning Research, 13, 3133-3176.

  K.H. Brodersen, C.S. Ong, J.M. Buhmann, & K.E. Stephan (2010). The balanced
  accuracy and its posterior distribution. ICPR, 3121-3124.
}
\author{
Kay H. Brodersen, ETH Zurich
}
